requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyascii==0.3.2
anyio==4.3.0
apparmor==3.1.7
appdirs==1.4.4
application-utility==1.3.2
argcomplete==2.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciidoc==10.2.0
asgiref==3.7.2
astroid==3.1.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
autocommand==2.2.2
Babel==2.12.1
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
black==23.12.1
bleach==6.1.0
btrfsutil==6.7.1
CacheControl==0.14.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
codespell==2.2.6
colorama==0.4.6
comm==0.2.1
configobj==5.0.8
coverage==7.4.1
cryptography==41.0.7
cssselect==1.2.0
cssselect2==0.7.0
cupshelpers==1.0
Cython==3.0.9
dbus-python==1.3.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
Django==4.2.11
django-extensions==3.2.3
docopt==0.6.2
duplicity==2.1.4
entrypoints==0.4
ewmh==0.1.6
executing==2.0.0
fasteners==0.19
fastjsonschema==2.19.1
filelock==3.13.1
flake8==7.0.0
fqdn==1.5.1
frozenlist==1.4.1
future==0.18.3
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
gufw==24.4.0
h11==0.14.0
httpcore==1.0.2
httpie==3.2.1
httplib2==0.22.0
httpx==0.26.0
idna==3.6
importlib_metadata==5.1.0
inflect==7.0.0
iotop==0.6
ipykernel==6.29.3
ipython==8.22.1
isodate==0.6.1
isoduration==20.11.0
isort==5.13.2
jaraco.context==4.3.0
jaraco.functools==4.0.0
jaraco.text==3.12.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.19
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.13.0
jupyterlab==4.1.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
layoutswitcherlib==0.8.37
lazy-object-proxy==1.10.0
lensfun==0.3.4
LibAppArmor==3.1.7
lit==16.0.6.dev0
lockfile==0.12.2
lxml==4.9.2
Mako==1.3.2
manjaro-sdk==0.1.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
material-color-utilities-python==0.1.5
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
meson==1.3.2
mistune==3.0.2
more-itertools==10.1.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.1.1
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
nest_asyncio==1.5.9
netsnmp-python==1.0a1
nftables==0.1
notebook==7.1.1
notebook_shim==0.2.4
npyscreen==4.10.5
numpy==1.25.0
oauth2client==4.1.3
ordered-set==4.1.0
overrides==7.7.0
packaging==23.2
pacman_mirrors==4.24.0
pandas==2.0.3
pandocfilters==1.5.1
paramiko==2.11.1
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.1.0
ply==3.11
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pulsectl==23.5.2
pure-eval==0.2.2
pwquality==1.4.5
pyaml==23.9.0
pyasn1==0.5.0
pyasn1_modules==0.3.0
pycairo==1.26.0
pycodestyle==2.11.1
pycparser==2.21
pycups==2.0.1
pycurl==7.45.2
pydantic==2.6.3
pydantic_core==2.16.3
PyDrive2==1.17.0
pyflakes==3.2.0
Pygments==2.17.2
PyGObject==3.46.0
pyliblo==0.10.0
pylint==3.1.0
PyNaCl==1.4.0
pyOpenSSL==24.0.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-sip==12.13.0
pyrsistent==0.19.3
pyserial==3.5
PySide6==6.6.2
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-xlib==0.33
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
ranger-fm==1.9.3
rdflib==6.2.0
referencing==0.33.0
regex==2023.12.25
reportlab==4.1.0
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rfc3339_validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.17.1
rsa==4.9
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
scour==0.38.2
Send2Trash==1.8.2
setproctitle==1.3.3
shiboken2==5.15.12
shiboken6==6.6.2
shiboken6-generator==6.6.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.3
stack-data==0.6.3
svglib==1.5.1
systemd-python==235
TBB==0.2
testpath==0.6.0
tinycss2==1.2.1
tldextract==5.1.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.4
tornado==6.3.2
tqdm==4.66.2
traitlets==5.14.1
trove-classifiers==2024.3.3
typing_extensions==4.10.0
tzdata==2023.3
uc-micro-py==1.0.3
ueberzug==18.2.1
ufw==0.36.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.18
validate==5.0.8
validate-pyproject==0.13.post1.dev0+gb752273.d20230520
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
wrapt==1.16.0
Yapsy==1.12.2
yarl==1.9.4
youtube_dl==2021.12.17
zipp==3.17.0
zstandard==0.22.0
